refactor: SectionService의 패키지를 ..service 에서 ..application 으로 이
